fdinfo: Extract FdinfoEntry from dump_one_file
To support SCMs we'll need to receive them into criu task
(see the SCM patch for details), then dump the received
file as if it was in the dumpee. Then the info about received
descriptor will be written into packet entry.

For this we'll need to perform all the regular file dumping
 code BUT not write the FdinfoEntry into image, so shuffle
the code for that.

Signed-off-by: Pavel Emelyanov <xemul@virtuozzo.com>